# PeopleCounter_Manager/docs/swagger.yaml
basePath: /
definitions:
  PeopleCounter_Manager_internal_models.APIStatsResponse:
    properties:
      counter_id:
        type: integer
      data:
        items:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.CounterStatsRecord'
        type: array
      period:
        type: string
    type: object
  PeopleCounter_Manager_internal_models.Counter:
    properties:
      created_at:
        type: string
      description:
        type: string
      groups:
        items:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.LinesGroup'
        type: array
      id:
        type: integer
      lines:
        items:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.Line'
        type: array
      name:
        type: string
      service_id:
        type: integer
      url:
        type: string
      vid_stride:
        type: integer
    type: object
  PeopleCounter_Manager_internal_models.CounterService:
    properties:
      api_key:
        type: string
      api_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  PeopleCounter_Manager_internal_models.CounterStatsRecord:
    properties:
      group_id:
        type: integer
      passes:
        type: integer
      period:
        type: string
    type: object
  PeopleCounter_Manager_internal_models.Line:
    properties:
      end:
        description: '[x, y]'
        items:
          type: integer
        type: array
      group_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      start:
        description: '[x, y]'
        items:
          type: integer
        type: array
    type: object
  PeopleCounter_Manager_internal_models.LinesGroup:
    properties:
      description:
        type: string
      id:
        type: integer
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API для управления счетчиками посетителей и сбора статистики
  termsOfService: http://swagger.io/terms/
  title: People Counter Manager API
  version: "1.0"
paths:
  /api/counters:
    get:
      consumes:
      - application/json
      description: Возвращает все камеры (админам) или только назначенные пользователю
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/PeopleCounter_Manager_internal_models.Counter'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить список счетчиков
      tags:
      - Counters
    post:
      consumes:
      - application/json
      description: Сохраняет камеру в БД и отправляет её конфигурацию (линии, группы)
        на привязанный воркер.
      parameters:
      - description: Конфигурация камеры (url, lines, groups, service_id)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.Counter'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.Counter'
        "400":
          description: Неверный формат JSON
          schema:
            type: string
        "500":
          description: Ошибка сервера или отказ воркера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать новый счетчик (камеру)
      tags:
      - Counters
  /api/counters/{id}:
    delete:
      description: Удаляет камеру из БД и отправляет команду на остановку потока на
        воркер.
      parameters:
      - description: ID камеры
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление
        "400":
          description: Неверный ID
          schema:
            type: string
        "404":
          description: Камера не найдена в БД
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить счетчик (камеру)
      tags:
      - Counters
    put:
      consumes:
      - application/json
      description: Обновляет конфигурацию камеры (линии, группы, URL) на лету. Изменения
        применяются на воркере без остановки сервиса.
      parameters:
      - description: ID камеры
        in: path
        name: id
        required: true
        type: integer
      - description: Полная обновленная конфигурация
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.Counter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.Counter'
        "400":
          description: Неверный ID или JSON
          schema:
            type: string
        "500":
          description: Ошибка сервера или отказ воркера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить счетчик (камеру)
      tags:
      - Counters
  /api/counters/{id}/stats:
    get:
      description: Возвращает агрегированную историю проходов для камеры, сгруппированную
        по указанному периоду и ID группы.
      parameters:
      - description: ID камеры
        in: path
        name: id
        required: true
        type: integer
      - description: Период агрегации (hour, day, week, month, year). По умолчанию
          'day'
        in: query
        name: period
        type: string
      - description: 'Начало диапазона (RFC3339, например: 2026-03-01T00:00:00Z)'
        in: query
        name: date_start
        type: string
      - description: 'Конец диапазона (RFC3339, например: 2026-03-31T23:59:59Z)'
        in: query
        name: date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.APIStatsResponse'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка формирования статистики
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить статистику счетчика
      tags:
      - Counters
  /api/counters/{id}/stream:
    get:
      description: Проксирует живой MJPEG видеопоток с разметкой от воркера к клиенту.
        Поддерживает токен в параметре запроса (?jwt=token).
      parameters:
      - description: ID камеры
        in: path
        name: id
        required: true
        type: integer
      - description: JWT Token (альтернатива заголовку Authorization)
        in: query
        name: jwt
        type: string
      produces:
      - multipart/x-mixed-replace
      responses:
        "200":
          description: MJPEG Stream
          schema:
            type: string
        "400":
          description: Неверный ID
          schema:
            type: string
        "502":
          description: Не удалось подключиться к видеопотоку
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить MJPEG видеопоток
      tags:
      - Counters
...
  /api/services/{id}:
    delete:
      description: Удаляет Python-воркер из БД. Связанные с ним камеры также будут
        каскадно удалены.
      parameters:
      - description: ID сервиса
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление
        "400":
          description: Неверный ID
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить воркер (сервис)
      tags:
      - Services
    get:
      description: Возвращает данные одного Python-воркера по его ID (APIKey скрыт
        в ответе).
      parameters:
      - description: ID сервиса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.CounterService'
        "400":
          description: Неверный ID
          schema:
            type: string
        "404":
          description: Сервис не найден
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить информацию о воркере
      tags:
      - Services
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего Python-воркера по его ID.
      parameters:
      - description: ID сервиса
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные сервиса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.CounterService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.CounterService'
        "400":
          description: Неверный ID или формат JSON
          schema:
            type: string
        "500":
          description: Ошибка сервера или сервис не найден
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить воркер (сервис)
      tags:
      - Services
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"